# Conflicts:
#	sistema_seguridad/src/sistema_seguridad/Imagen.java

+++ sistema_seguridad/src/sistema_seguridad/Reloj.java
@@ -17,7 +17,7 @@ import java.util.logging.Level;
 public class Reloj implements Runnable {
     public static Reloj instance = null;
     public static int velocidad = 1000;
-    private long momentoActual;
+    private int momentoActual;
     
 
     public int getVelocidad() {
@@ -28,30 +28,24 @@ public class Reloj implements Runnable {
         this.velocidad = velocidad;
     }
     
-    
-    
     public Reloj(){
         this.momentoActual = 0;
     }
 
-    public long getMomentoActual() {
+    public int getMomentoActual() {
         return momentoActual;
     }
 
-    public void setMomentoActual(long momentoActual) {
+    public void setMomentoActual(int momentoActual) {
         this.momentoActual = momentoActual;
-    }
-    
+    }    
     
     public static Reloj getInstance(){
         if (instance ==null){
             instance = new Reloj();
         }
         return instance;
-    }
-    
-    
-    
+    }  
 
     @Override
     public void run() {
@@ -68,20 +62,10 @@ public class Reloj implements Runnable {
                 
                 Logger.getInstancia().log("cambio el momento");
                 
-                
-                
             } catch (IOException ex) {
                 java.util.logging.Logger.getLogger(Reloj.class.getName()).log(Level.SEVERE, null, ex);
 //                Logger.getLogger(Reloj.class.getName()).log(Level.SEVERE, null, ex);
             }
-            
-            
-            
         }
-        
-        
-        
     }
-    
-    
  }
